Define crop window relative to the image's display window
Previously, it was defined relative to the data window, which is inaccurate
when actually making use of OpenEXR's display<->data distinction.

+++ include/tev/ImageViewer.h
@@ -259,7 +259,7 @@ private:
     size_t mDraggedImageButtonId;
 
     nanogui::Vector2i mCroppingStartCoordinates;
-    nanogui::Vector2f mDraggingStartPosition;
+    nanogui::Vector2i mDraggingStartPosition;
 
     size_t mClipboardIndex = 0;
